+++ components/ui/shared/slider.tsx
@@ -15,9 +15,8 @@ import * as React from "react";
 import * as SliderPrimitive from "@radix-ui/react-slider";
 import { cn } from "@/lib/utils";
 
-interface SliderProps extends React.ComponentPropsWithoutRef<
-  typeof SliderPrimitive.Root
-> {
+interface SliderProps
+  extends React.ComponentPropsWithoutRef<typeof SliderPrimitive.Root> {
   thumbCount?: number;
 }